room-expander/pyproject.toml
[tool.poetry]
name = "room-expander"
version = "0.1.0"
description = ""
authors = ["tgscan.xyz@outlook.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.29.3"
loguru = "^0.7.0"
jsonpickle = "^3.0.2"
tortoise-orm = "^0.20.0"
toolz = "^0.12.0"
python-dotenv = "^1.0.0"
rich = "^13.5.2"
sqlalchemy = "^2.0.22"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
